Performance Fixes

+++ crates/assistd/src/chat/app/tests.rs
@@ -106,7 +106,7 @@ fn status(event: Event) -> ChatEvent {
 /// Accept one dialog connection, read its request line, then stream
 /// `events` back. The delay before the first write gives the query
 /// driver time to observe its closed writer channel while nothing is
-/// readable, which is the state the pre-fix driver parked in forever.
+/// readable, so a driver that parks in that state hangs the test.
 async fn mock_daemon(
     socket: std::path::PathBuf,
     events: Vec<Event>,
@@ -298,7 +298,11 @@ fn mouse_wheel_scrolls_by_a_fixed_step() {
         (MouseEventKind::ScrollDown, MOUSE_WHEEL_STEP),
     ] {
         app.on_mouse(wheel(kind));
-        assert_eq!(app.output.scroll_offset(), expected, "after {kind:?}");
+        assert_eq!(
+            app.output.scroll_offset(),
+            usize::from(expected),
+            "after {kind:?}"
+        );
     }
 }
